Extract storages.py and renderers.py from vips.py
In vips.py: TmsTile used to combine storage with rendering
images. This has now been split into two different classes that are
passed into TmsPyramid or TmsTiles. This means that TmsTile is
removed and the TmsTiles needs to know about its current resolution.

We have temporarily changed the behaviour of TmsPyramid so it no
longer writes into nested subdirectories. This will change once we
implement a NestedFileStorage.