Correct for GDAL warping's tendency to eat pixels when reprojecting.
GDAL sometimes reprojects from a whole world image into a partial
world image, due to rounding errors. Therefore, we must upsample the
image to fit whole tiles, even if this makes the extents of the
image go PAST the full world. Otherwise, there will be a missing
pixels on the eastern border and doing an alignment would cause a gap
in the world when crossing the dateline.