Move zoom_level offset handling to MbtilesStorage.
The old implementation, where we had the tiles view compute the
offset, was architecturally cleaner but slower because SQLite had to
scan the entire map table before it could locate a single tile.

Instead, there are only two lookups now, directly from the index,
because the zoom_level for the map table is already offset when we
write the file.