Adjusted the number of pixels when upsampling to not lose any data.
When upsampling, if the scaling ratios cause the number of pixels of the
output image to be a fraction of a pixel, we round up so we don't lose any
data.

Before, we were rounding the number of pixel to the closest integer, but that
means that we could lose up to half a pixel worth of data.

This shouldn't have much of an impact on the output image, especially when
using nearest neighbour.