vips.py: Proper band sizes, copy buffers in ReadAsArray to prevent corruption
The band's X and Y size was not referring to the dataset, so if you got a
handle on the band and then scaled the dataset, the X and Y sizes were not
changed and were thus wrong. Always report the X and Y sizes from the dataset.

There was a bug in the band and dataset ReadAsArray methods where if you tried
to read from the buffer, the values could be modified because the ndarray
returned was referring to the direct buffer in memory. Copying the buffer
before returning the ndarray is the correct thing to do.