gdal.py: Fix a tile alignment problem.
Some times, when upscaling and slicing a partial world, the bottom TMS extents,
which is used to determine the Y position of tiles, would be wrongfully
rounded down, thus rendering the tiles exactly 1 tile downward from where
they should be on the map.

Rounding the calculations of the TMS extents before converting the floating
point values into integers fixes the problem.