Warp bluemarble.tif to web mercator
It was in EPSG:3785, which used to be the default in gdal2mbtiles
until it was changed by
https://github.com/ecometrica/gdal2mbtiles/commit/822cae9de8088d49f401104e6a04f130cd6c7eb2

The tests were never updated, so `test_get_spatial_reference` was
failing. I think it makes more sense to warp the test files
to the right default EPSG_WEB_MERCATOR instead of just testing
that the existing file was in EPSG:3785.

The other test files will be reprojected too.