Allow for either stretching or shrinking on initial resample
When resampling to the native resolution after initial warp,
we should not always assuming that we are going to upsample
to the "native" resolution, because of the floating point
error tolerance in GetNativeResolutoin.

This now looks at the scaling ratios and calls
either `stretch` or `shrink` based on the values.