fix: Use the major Circumference function if the spatial reference is 3857.
When initiating a spatial reference using ImportProj4(raster.crs.to_string())
where crs.to_string is a rasterio Dataset method,
the spatial reference GetInvFlattening() returns a 0 and causes a division by zero
error.
The 3857 is a projection from a sphere(not a spheroid or ellipsoid), so the
Major and Minor should be equal.